Se sacan los métodos restantes de app.js y pasan a controladores
